+++ src/components/Modal.tsx
@@ -1,16 +1,18 @@
 import { ReactNode, PropsWithChildren } from 'react'
 import * as rb from 'react-bootstrap'
 import { useTranslation } from 'react-i18next'
-import styles from './Modal.module.css'
 import Sprite from './Sprite'
+import styles from './Modal.module.css'
 
-type BaseModalProps = {
-  isShown: boolean
-  title: ReactNode | string
-  onCancel: () => void
-  backdrop?: rb.ModalProps['backdrop']
-  size?: rb.ModalProps['size']
-}
+type BaseModalProps = Pick<rb.ModalProps, 'className' | 'backdrop' | 'size'> &
+  Pick<rb.ModalHeaderProps, 'closeButton'> & {
+    isShown: boolean
+    title: ReactNode | string
+    onCancel: () => void
+    headerClassName?: rb.ModalHeaderProps['className']
+    titleClassName?: rb.ModalTitleProps['className']
+  }
+
 const BaseModal = ({
   isShown,
   title,
@@ -18,6 +20,10 @@ const BaseModal = ({
   onCancel,
   size,
   backdrop = 'static',
+  closeButton = false,
+  className = styles.modal,
+  headerClassName = styles.modalHeader,
+  titleClassName = styles.modalTitle,
 }: PropsWithChildren<BaseModalProps>) => {
   return (
     <rb.Modal
@@ -29,10 +35,10 @@ const BaseModal = ({
       animation={true}
       backdrop={backdrop}
       size={size}
-      className={styles.modal}
+      className={className}
     >
-      <rb.Modal.Header className={styles['modal-header']}>
-        <rb.Modal.Title className={styles['modal-title']}>{title}</rb.Modal.Title>
+      <rb.Modal.Header className={headerClassName} closeButton={closeButton}>
+        <rb.Modal.Title className={titleClassName}>{title}</rb.Modal.Title>
       </rb.Modal.Header>
       {children}
     </rb.Modal>
@@ -53,8 +59,8 @@ const InfoModal = ({
 }: PropsWithChildren<InfoModalProps>) => {
   return (
     <BaseModal {...baseModalProps} onCancel={onCancel} backdrop={true}>
-      <rb.Modal.Body className={styles['modal-body']}>{children}</rb.Modal.Body>
-      <rb.Modal.Footer className={styles['modal-footer']}>
+      <rb.Modal.Body className={styles.modalBody}>{children}</rb.Modal.Body>
+      <rb.Modal.Footer className={styles.modalFooter}>
         <rb.Button variant="outline-dark" onClick={() => onSubmit()}>
           {submitButtonText}
         </rb.Button>
@@ -63,17 +69,24 @@ const InfoModal = ({
   )
 }
 
-export type ConfirmModalProps = BaseModalProps & {
+export type ConfirmModalProps = Omit<BaseModalProps, 'closeButton'> & {
   onConfirm: () => void
+  disabled?: boolean
 }
 
-const ConfirmModal = ({ children, onCancel, onConfirm, ...baseModalProps }: PropsWithChildren<ConfirmModalProps>) => {
+const ConfirmModal = ({
+  children,
+  onCancel,
+  onConfirm,
+  disabled = false,
+  ...baseModalProps
+}: PropsWithChildren<ConfirmModalProps>) => {
   const { t } = useTranslation()
 
   return (
     <BaseModal {...baseModalProps} onCancel={onCancel}>
-      <rb.Modal.Body className={styles['modal-body']}>{children}</rb.Modal.Body>
-      <rb.Modal.Footer className={styles['modal-footer']}>
+      <rb.Modal.Body className={styles.modalBody}>{children}</rb.Modal.Body>
+      <rb.Modal.Footer className={styles.modalFooter}>
         <rb.Button
           variant="outline-dark"
           onClick={() => onCancel()}
@@ -82,7 +95,7 @@ const ConfirmModal = ({ children, onCancel, onConfirm, ...baseModalProps }: Prop
           <Sprite symbol="cancel" width="26" height="26" />
           <div>{t('modal.confirm_button_reject')}</div>
         </rb.Button>
-        <rb.Button variant="outline-dark" onClick={() => onConfirm()}>
+        <rb.Button variant="outline-dark" onClick={() => onConfirm()} disabled={disabled}>
           {t('modal.confirm_button_accept')}
         </rb.Button>
       </rb.Modal.Footer>
@@ -90,4 +103,4 @@ const ConfirmModal = ({ children, onCancel, onConfirm, ...baseModalProps }: Prop
   )
 }
 
-export { InfoModal, ConfirmModal }
+export { BaseModal, InfoModal, ConfirmModal }