+++ src/components/Send/SendForm.tsx
@@ -1,6 +1,6 @@
-import { useState, useMemo } from 'react'
+import { useState, useMemo, useEffect } from 'react'
 import { Trans, useTranslation } from 'react-i18next'
-import { Field, Formik, FormikErrors, FormikProps } from 'formik'
+import { Field, Formik, FormikErrors, FormikProps, useField } from 'formik'
 import * as rb from 'react-bootstrap'
 import * as Api from '../../libs/JmWalletApi'
 import ToggleSwitch from '../ToggleSwitch'
@@ -26,7 +26,7 @@ import {
   isValidNumCollaborators,
 } from './helpers'
 import { AccountBalanceSummary } from '../../context/BalanceSummary'
-import { WalletInfo } from '../../context/WalletContext'
+import { WalletInfo, CurrentWallet } from '../../context/WalletContext'
 import { useSettings } from '../../context/SettingsContext'
 import styles from './SendForm.module.css'
 import { TxFeeInputField, validateTxFee } from '../settings/TxFeeInputField'
@@ -67,7 +67,7 @@ function CollaborativeTransactionOptions({
     amount:
       selectedAmount?.isSweep && sourceJarBalance
         ? sourceJarBalance.calculatedAvailableBalanceInSats
-        : selectedAmount?.value ?? null,
+        : (selectedAmount?.value ?? null),
     numCollaborators: selectedNumCollaborators ?? null,
     isCoinjoin: true,
   })
@@ -134,8 +134,8 @@ function CollaborativeTransactionOptions({
           numCollaborators={selectedNumCollaborators ?? null}
           amount={
             selectedAmount?.isSweep
-              ? sourceJarBalance?.calculatedAvailableBalanceInSats ?? null
-              : selectedAmount?.value ?? null
+              ? (sourceJarBalance?.calculatedAvailableBalanceInSats ?? null)
+              : (selectedAmount?.value ?? null)
           }
           onClick={() => {
             setActiveFeeConfigModalSection('cj_fee')
@@ -221,6 +221,7 @@ interface InnerSendFormProps {
   className?: string
   isLoading: boolean
   walletInfo?: WalletInfo
+  wallet: CurrentWallet
   loadNewWalletAddress: (props: { signal: AbortSignal; jarIndex: JarIndex }) => Promise<Api.BitcoinAddress>
   minNumCollaborators: number
   feeConfigValues?: FeeValues
@@ -233,6 +234,7 @@ const InnerSendForm = ({
   className,
   isLoading,
   walletInfo,
+  wallet,
   loadNewWalletAddress,
   minNumCollaborators,
   feeConfigValues,
@@ -241,6 +243,9 @@ const InnerSendForm = ({
 }: InnerSendFormProps) => {
   const { t } = useTranslation()
   const serviceInfo = useServiceInfo()
+  const amountField = useField<AmountValue>('amount')
+  const amountMeta = amountField[1]
+  const amountHelper = amountField[2]
 
   const jarBalances = useMemo(() => {
     if (!walletInfo) return []
@@ -259,11 +264,25 @@ const InnerSendForm = ({
     return buildCoinjoinRequirementSummary(sourceJarUtxos)
   }, [sourceJarUtxos])
 
-  const sourceJarBalance =
-    props.values.sourceJarIndex !== undefined ? jarBalances[props.values.sourceJarIndex] : undefined
+  const sourceJarBalance = useMemo(
+    () => (props.values.sourceJarIndex !== undefined ? jarBalances[props.values.sourceJarIndex] : undefined),
+    [jarBalances, props.values.sourceJarIndex],
+  )
 
   const showCoinjoinPreconditionViolationAlert =
     !isLoading && !disabled && props.values.isCoinJoin && sourceJarCoinjoinPreconditionSummary?.isFulfilled === false
+
+  //Effect to change the field value whenever the sourceJarBalance changes (sourceJarBalance will change when quick freeze/unfreeze is performed or different source jar is selected)
+  useEffect(() => {
+    if (!sourceJarBalance) return
+    amountHelper.setValue(
+      amountMeta.initialValue || {
+        value: null,
+        isSweep: false,
+      },
+      true,
+    )
+  }, [sourceJarBalance, amountHelper, amountMeta.initialValue])
 
   return (
     <>
@@ -272,6 +291,7 @@ const InnerSendForm = ({
           name="sourceJarIndex"
           label={t('send.label_source_jar')}
           walletInfo={walletInfo}
+          wallet={wallet}
           isLoading={isLoading}
           disabled={disabled}
           variant={showCoinjoinPreconditionViolationAlert ? 'warning' : 'default'}
@@ -375,6 +395,7 @@ type SendFormProps = Omit<InnerSendFormProps, 'props' | 'className'> & {
   onSubmit: (values: SendFormValues) => Promise<void>
   formRef?: React.Ref<FormikProps<SendFormValues>>
   blurred?: boolean
+  wallet: CurrentWallet
 }
 
 export const SendForm = ({
@@ -383,6 +404,7 @@ export const SendForm = ({
   formRef,
   blurred = false,
   walletInfo,
+  wallet,
   minNumCollaborators,
   ...innerProps
 }: SendFormProps) => {
@@ -446,6 +468,7 @@ export const SendForm = ({
             props={props}
             className={blurred ? styles.blurred : undefined}
             walletInfo={walletInfo}
+            wallet={wallet}
             minNumCollaborators={minNumCollaborators}
             {...innerProps}
           />