+++ src/components/Send/index.tsx
@@ -4,6 +4,7 @@ import { FormikProps } from 'formik'
 import { useTranslation } from 'react-i18next'
 import * as rb from 'react-bootstrap'
 import * as Api from '../../libs/JmWalletApi'
+import { isDevMode } from '../../constants/debugFeatures'
 import PageTitle from '../PageTitle'
 import Sprite from '../Sprite'
 import { SendForm, SendFormValues } from './SendForm'
@@ -18,13 +19,15 @@ import { useLoadConfigValue } from '../../context/ServiceConfigContext'
 import { useWaitForUtxosToBeSpent } from '../../hooks/WaitForUtxosToBeSpent'
 import { routes } from '../../constants/routes'
 import { JM_MINIMUM_MAKERS_DEFAULT } from '../../constants/config'
-
 import { initialNumCollaborators } from './helpers'
 
 const INITIAL_DESTINATION = null
 const INITIAL_SOURCE_JAR_INDEX = null
 const INITIAL_AMOUNT = null
 const INITIAL_IS_COINJOIN = true
+
+// set the default to one collaborat
+const DEV_INITIAL_NUM_COLLABORATORS_INPUT = 1
 
 type MaxFeeConfigMissingAlertProps = {
   onSuccess: () => void
@@ -98,7 +101,10 @@ export default function Send({ wallet }: SendProps) {
   const [alert, setAlert] = useState<SimpleAlert>()
   const [isSending, setIsSending] = useState(false)
   const [minNumCollaborators, setMinNumCollaborators] = useState(JM_MINIMUM_MAKERS_DEFAULT)
-  const initNumCollaborators = useMemo(() => initialNumCollaborators(minNumCollaborators), [minNumCollaborators])
+  const initNumCollaborators = useMemo(
+    () => (isDevMode() ? DEV_INITIAL_NUM_COLLABORATORS_INPUT : initialNumCollaborators(minNumCollaborators)),
+    [minNumCollaborators],
+  )
 
   const [feeConfigValues, reloadFeeConfigValues] = useFeeConfigValues()
   const maxFeesConfigMissing = useMemo(
@@ -201,15 +207,13 @@ export default function Send({ wallet }: SendProps) {
         setAlert({ variant: 'danger', message })
       })
 
-      const loadingWalletInfoAndUtxos = reloadCurrentWalletInfo
-        .reloadUtxos({ signal: abortCtrl.signal })
-        .catch((err) => {
-          if (abortCtrl.signal.aborted) return
-          const message = t('global.errors.error_loading_wallet_failed', {
-            reason: err.message || t('global.errors.reason_unknown'),
-          })
-          setAlert({ variant: 'danger', message })
+      const loadingWalletInfoAndUtxos = reloadCurrentWalletInfo.reloadAll({ signal: abortCtrl.signal }).catch((err) => {
+        if (abortCtrl.signal.aborted) return
+        const message = t('global.errors.error_loading_wallet_failed', {
+          reason: err.message || t('global.errors.reason_unknown'),
         })
+        setAlert({ variant: 'danger', message })
+      })
 
       const loadingMinimumMakerConfig = loadConfigValue({
         signal: abortCtrl.signal,
@@ -433,7 +437,7 @@ export default function Send({ wallet }: SendProps) {
               </div>
               <rb.Alert variant="success" className="d-flex align-items-center">
                 {t('send.text_coinjoin_already_running')}
-                <Sprite className="ms-auto" symbol="joining" width="20" height="20" />
+                <Sprite className="ms-auto" symbol="mixed" width="24" height="24" />
               </rb.Alert>
               <rb.Button
                 variant="none"
@@ -480,6 +484,7 @@ export default function Send({ wallet }: SendProps) {
         disabled={isOperationDisabled}
         isLoading={isLoading}
         walletInfo={walletInfo}
+        wallet={wallet}
         minNumCollaborators={minNumCollaborators}
         loadNewWalletAddress={loadNewWalletAddress}
         feeConfigValues={feeConfigValues}
@@ -516,6 +521,9 @@ export default function Send({ wallet }: SendProps) {
             isCoinjoin: showConfirmSendModal.isCoinJoin,
             numCollaborators: showConfirmSendModal.numCollaborators!,
             feeConfigValues: { ...feeConfigValues, tx_fees: showConfirmSendModal.txFee },
+            availableUtxos: (walletInfo?.utxosByJar[showConfirmSendModal.sourceJarIndex!] || [])
+              .filter((utxo) => !utxo.frozen)
+              .sort((a, b) => a.confirmations - b.confirmations),
           }}
         />
       )}