+++ src/components/Settings.tsx
@@ -16,7 +16,7 @@ import languages from '../i18n/languages'
 import styles from './Settings.module.css'
 import SeedModal from './settings/SeedModal'
 import FeeConfigModal from './settings/FeeConfigModal'
-import { isDebugFeatureEnabled, isDevMode } from '../constants/debugFeatures'
+import { isDevMode } from '../constants/debugFeatures'
 import { isFeatureEnabled } from '../constants/features'
 import { CurrentWallet } from '../context/WalletContext'
 
@@ -213,16 +213,10 @@ export default function Settings({ wallet, stopWallet }: SettingsProps) {
             {t('settings.button_switch_wallet')}
           </Link>
 
-          {serviceInfo && isFeatureEnabled('rescanChain', serviceInfo) && isDebugFeatureEnabled('rescanChainPage') && (
-            <Link
-              to={routes.rescanChain}
-              className={`btn btn-outline-dark ${styles['settings-btn']} position-relative`}
-            >
+          {serviceInfo && isFeatureEnabled('rescanChain', serviceInfo) && (
+            <Link to={routes.rescanChain} className={`btn btn-outline-dark ${styles['settings-btn']}`}>
               <Sprite symbol="block" width="24" height="24" />
-              Rescan chain
-              <span className="position-absolute top-50 start-0 translate-middle badge rounded-pill bg-warning">
-                dev
-              </span>
+              {t('settings.rescan_chain')}
             </Link>
           )}
 
@@ -254,12 +248,6 @@ export default function Settings({ wallet, stopWallet }: SettingsProps) {
             <div className="d-flex align-items-center">
               <Sprite symbol="telegram" width="24" height="24" className="me-2 p-1" />
               {t('settings.telegram')}
-            </div>
-          </a>
-          <a href="https://twitter.com/joinmarket" target="_blank" rel="noopener noreferrer" className="link-dark">
-            <div className="d-flex align-items-center">
-              <Sprite symbol="twitter" width="24" height="24" className="me-2 p-1" />
-              {t('settings.jm_twitter')}
             </div>
           </a>
         </div>